requirements.txt
streamlit 
google-generativeai
python-dotenv
langchain
PyPDF2
faiss-cpu
langchain_google_genai
pdf2image
langchain_community
numpy 
pandas
Requests
poppler-utils